For autocomplete: Swap from retrieving WordProcessor to retrieving autocomplete list
I found out that there is a bug in the retrieved WordProcessor. It appears that the retrieved WordProcessor does not actually retain its own method coding; rather, it retrieves the local code when its methods are called. This is clearly incorrect. Consequently,  rewrote testCasesAdvanced, testcasesClient, and testcasesServer to account for passing the List when doing autocomplete checking.