Minimize amount of networking required. Greatly speeds up test Z.
Prior to this change, the test cases would send the server a prefix, and the server would return the autocompleted list. Because there's so much networking involved in this, it was very very slow (to the tune of 1.5 minutes on the default wordsToAdd). By removing this and having the server immediately calculate all combinations of valid prefixes and returning that list, networking is necessary exponentially fewer times. Now runs to the tune of 150 ms.